requirements.txt
netaddr
paramiko
python-novaclient
